# services/chatbot/src/resources/crapi-openapi-spec.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OWASP crAPI API",
    "version" : "1-oas3"
  },
...
  "servers" : [ {
    "url" : "http://localhost:8888"
  } ],
  "paths" : {
...
    "/identity/api/v2/vehicle/vehicles" : {
      "get" : {
        "operationId" : "get_vehicles",
        "summary" : "Get user vehicles",
        "tags" : [ "Identity / Vehicle" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "object",
                    "required" : [ "id", "model", "owner", "pincode", "status", "uuid", "vehicleLocation", "vin", "year" ],
                    "properties" : {
                      "id" : {
                        "type" : "number"
                      },
                      "uuid" : {
                        "type" : "string"
                      },
                      "year" : {
                        "type" : "number"
                      },
                      "status" : {
                        "type" : "string"
                      },
                      "vin" : {
                        "type" : "string"
                      },
                      "pincode" : {
                        "type" : "string"
                      },
                      "owner" : {
                        "$ref" : "#/components/schemas/User"
                      },
                      "model" : {
                        "type" : "object",
                        "required" : [ "fuel_type", "id", "model", "vehicle_img", "vehiclecompany" ],
                        "properties" : {
                          "model" : {
                            "type" : "string"
                          },
                          "fuel_type" : {
                            "type" : "string"
                          },
                          "vehicle_img" : {
                            "type" : "string"
                          },
                          "id" : {
                            "type" : "number"
                          },
                          "vehiclecompany" : {
                            "type" : "object",
                            "required" : [ "id", "name" ],
                            "properties" : {
                              "id" : {
                                "type" : "number"
                              },
                              "name" : {
                                "type" : "string"
                              }
                            }
                          }
                        }
                      },
                      "vehicleLocation" : {
                        "type" : "object",
                        "required" : [ "id", "latitude", "longitude" ],
                        "properties" : {
                          "id" : {
                            "type" : "number"
                          },
                          "latitude" : {
                            "type" : "string"
                          },
                          "longitude" : {
                            "type" : "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "description" : ""
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CRAPIResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        }
      },
      "parameters" : [ ]
    },
    "/identity/api/v2/vehicle/verify_vehicle" : {
      "post" : {
        "operationId" : "verify_vehicle",
        "summary" : "Add the user vehicle",
        "tags" : [ "Identity / Vehicle" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "pincode", "vin" ],
                "properties" : {
                  "pincode" : {
                    "type" : "string",
                    "example" : "9896"
                  },
                  "vin" : {
                    "type" : "string",
                    "example" : "0IOJO38SMVL663989"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CRAPIResponse"
                }
              }
            },
            "description" : ""
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CRAPIResponse"
                }
              }
            },
            "description" : ""
          }
        }
      },
      "parameters" : [ ]
    },
    "/identity/api/v2/vehicle/{vehicleId}/location" : {
      "get" : {
        "operationId" : "get_location",
        "summary" : "Get Vehicle Location",
        "description" : "Get user's vehicle location",
        "tags" : [ "Identity / Vehicle" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "required" : [ "carId", "fullName", "vehicleLocation" ],
                  "properties" : {
                    "carId" : {
                      "type" : "string"
                    },
                    "fullName" : {
                      "type" : "string"
                    },
                    "vehicleLocation" : {
                      "type" : "object",
                      "required" : [ "id", "latitude", "longitude" ],
                      "properties" : {
                        "id" : {
                          "type" : "number"
                        },
                        "latitude" : {
                          "type" : "string"
                        },
                        "longitude" : {
                          "type" : "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "description" : ""
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CRAPIResponse"
                }
              }
            },
            "description" : "Invalid vehicle_id for User"
          }
        }
      },
      "parameters" : [ {
        "in" : "path",
        "name" : "vehicleId",
        "required" : true,
        "schema" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "1929186d-8b67-4163-a208-de52a41f7301"
        }
      } ]
    },
    "/identity/api/v2/vehicle/resend_email" : {
      "post" : {
        "operationId" : "vehicle_resend_email",
        "summary" : "Resend Vehicle Details Email",
        "description" : "Resend vehicles details to be added to the user dashboard",
        "tags" : [ "Identity / Vehicle" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CRAPIResponse"
                }
              }
            },
            "description" : "OK"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CRAPIResponse"
                }
              }
            },
            "description" : ""
          }
        }
      },
      "parameters" : [ ]
    },
...
  },
  "components" : {
    "schemas" : {
...
      "User" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "number" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "required" : [ "email" ]
      },
...
      "CRAPIResponse" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}